+++ src/components/admin/reservations/list/DateFilter.tsx
@@ -1,23 +1,23 @@
-import React from 'react';
-import { Calendar, Funnel, X } from 'lucide-react';
-import { Button } from '@/components/ui/button';
+import React from "react";
+import { Calendar, Funnel, X } from "lucide-react";
+import { Button } from "@/components/ui/button";
 import {
   Popover,
   PopoverContent,
   PopoverTrigger,
-} from '@/components/ui/popover';
-import { Calendar as CalendarComponent } from '@/components/ui/calendar';
+} from "@/components/ui/popover";
+import { Calendar as CalendarComponent } from "@/components/ui/calendar";
 import {
   Select,
   SelectContent,
   SelectItem,
   SelectTrigger,
   SelectValue,
-} from '@/components/ui/select';
+} from "@/components/ui/select";
 
-type DateFilterMode = 'specific' | 'range';
+type DateFilterMode = "specific" | "range";
 
-type DateFilterValue = {
+export type DateFilterValue = {
   mode: DateFilterMode;
   specificDate?: Date;
   startDate?: Date;
@@ -31,7 +31,7 @@ type DateFilterProps = {
 };
 
 export const parseDateString = (dateString: string): Date => {
-  const [year, month, day] = dateString.split('-').map(Number);
+  const [year, month, day] = dateString.split("-").map(Number);
   return new Date(year, month - 1, day, 0, 0, 0, 0);
 };
 
@@ -41,7 +41,11 @@ export const normalizeDate = (date: Date): Date => {
   return normalized;
 };
 
-export default function DateFilter({ value, onChange, onClear }: DateFilterProps) {
+export default function DateFilter({
+  value,
+  onChange,
+  onClear,
+}: DateFilterProps) {
   const handleModeChange = (newMode: DateFilterMode) => {
     onChange({
       mode: newMode,
@@ -52,41 +56,29 @@ export default function DateFilter({ value, onChange, onClear }: DateFilterProps
   };
 
   const handleSpecificDateChange = (date: Date | undefined) => {
-    onChange({
-      ...value,
-      mode: 'specific',
-      specificDate: date,
-    });
+    onChange({ ...value, mode: "specific", specificDate: date });
   };
 
   const handleStartDateChange = (date: Date | undefined) => {
-    onChange({
-      ...value,
-      mode: 'range',
-      startDate: date,
-    });
+    onChange({ ...value, mode: "range", startDate: date });
   };
 
   const handleEndDateChange = (date: Date | undefined) => {
-    onChange({
-      ...value,
-      mode: 'range',
-      endDate: date,
-    });
+    onChange({ ...value, mode: "range", endDate: date });
   };
 
   const formatDate = (date: Date | undefined, placeholder: string) => {
     if (!date) return placeholder;
-    return date.toLocaleDateString('fr-CA', {
-      year: 'numeric',
-      month: 'long',
-      day: 'numeric',
+    return date.toLocaleDateString("fr-CA", {
+      year: "numeric",
+      month: "long",
+      day: "numeric",
     });
   };
 
   const hasActiveFilter =
-    (value.mode === 'specific' && value.specificDate) ||
-    (value.mode === 'range' && (value.startDate || value.endDate));
+    (value.mode === "specific" && value.specificDate) ||
+    (value.mode === "range" && (value.startDate || value.endDate));
 
   return (
     <div className="flex flex-col lg:flex-row gap-2 items-start lg:items-center">
@@ -105,7 +97,7 @@ export default function DateFilter({ value, onChange, onClear }: DateFilterProps
       </div>
 
       <div className="flex flex-col md:flex-row gap-2 w-full lg:w-auto">
-        {value.mode === 'specific' && (
+        {value.mode === "specific" && (
           <Popover>
             <PopoverTrigger asChild>
               <Button
@@ -113,7 +105,7 @@ export default function DateFilter({ value, onChange, onClear }: DateFilterProps
                 className="w-full md:w-[240px] justify-start text-left font-normal"
               >
                 <Calendar className="mr-2 h-4 w-4 text-cyan-500" />
-                {formatDate(value.specificDate, 'Sélectionner une date')}
+                {formatDate(value.specificDate, "Sélectionner une date")}
               </Button>
             </PopoverTrigger>
             <PopoverContent className="w-auto p-0" align="start">
@@ -127,7 +119,7 @@ export default function DateFilter({ value, onChange, onClear }: DateFilterProps
           </Popover>
         )}
 
-        {value.mode === 'range' && (
+        {value.mode === "range" && (
           <>
             <Popover>
               <PopoverTrigger asChild>
@@ -136,7 +128,7 @@ export default function DateFilter({ value, onChange, onClear }: DateFilterProps
                   className="w-full md:w-[200px] justify-start text-left font-normal"
                 >
                   <Calendar className="mr-2 h-4 w-4 text-cyan-500" />
-                  {formatDate(value.startDate, 'Date de début')}
+                  {formatDate(value.startDate, "Date de début")}
                 </Button>
               </PopoverTrigger>
               <PopoverContent className="w-auto p-0" align="start">
@@ -156,7 +148,7 @@ export default function DateFilter({ value, onChange, onClear }: DateFilterProps
                   className="w-full md:w-[200px] justify-start text-left font-normal"
                 >
                   <Calendar className="mr-2 h-4 w-4 text-cyan-500" />
-                  {formatDate(value.endDate, 'Date de fin')}
+                  {formatDate(value.endDate, "Date de fin")}
                 </Button>
               </PopoverTrigger>
               <PopoverContent className="w-auto p-0" align="start">
