+++ src/components/admin/reservations/list/Pagination.tsx
@@ -2,7 +2,12 @@
 
 import React, { useMemo } from "react";
 import { Button } from "@/components/ui/button";
-import { ChevronLeft, ChevronRight, ChevronsLeft, ChevronsRight } from "lucide-react";
+import {
+  ChevronLeft,
+  ChevronRight,
+  ChevronsLeft,
+  ChevronsRight,
+} from "lucide-react";
 import { cn } from "@/lib/utils";
 import { useTranslations } from "next-intl";
 
@@ -29,7 +34,10 @@ export default function PaginationControls({
   const isFirst = page <= 1;
   const isLast = page >= totalPages;
 
-  const range = useMemo(() => buildRange(page, totalPages, siblingCount), [page, totalPages, siblingCount]);
+  const range = useMemo(
+    () => buildRange(page, totalPages, siblingCount),
+    [page, totalPages, siblingCount],
+  );
 
   const startItem = totalItems === 0 ? 0 : (page - 1) * pageSize + 1;
   const endItem = Math.min(page * pageSize, totalItems);
@@ -37,18 +45,26 @@ export default function PaginationControls({
   const t = useTranslations();
 
   return (
-    <div className={cn("flex flex-col gap-3 sm:gap-4 pt-4 border-t", className)}>
+    <div
+      className={cn("flex flex-col gap-3 sm:gap-4 pt-4 border-t", className)}
+    >
       <div className="flex flex-col items-start justify-between gap-3 sm:flex-row sm:items-center">
         <div className="text-xs sm:text-sm text-muted-foreground">
           {totalItems > 0 ? (
             <>
               <span className="hidden sm:inline">
-                {t("admin.reservations.pagination.affichageDe")} <span className="font-medium text-foreground">{startItem}</span> {t("admin.reservations.pagination.a")}{" "}
-                <span className="font-medium text-foreground">{endItem}</span> {t("admin.reservations.pagination.sur")}{" "}
-                <span className="font-medium text-foreground">{totalItems}</span>
+                {t("admin.reservations.pagination.affichageDe")}{" "}
+                <span className="font-medium text-foreground">{startItem}</span>{" "}
+                {t("admin.reservations.pagination.a")}{" "}
+                <span className="font-medium text-foreground">{endItem}</span>{" "}
+                {t("admin.reservations.pagination.sur")}{" "}
+                <span className="font-medium text-foreground">
+                  {totalItems}
+                </span>
               </span>
               <span className="sm:hidden">
-                {startItem}-{endItem} {t("admin.reservations.pagination.sur")} {totalItems}
+                {startItem}-{endItem} {t("admin.reservations.pagination.sur")}{" "}
+                {totalItems}
               </span>
             </>
           ) : (
@@ -57,7 +73,10 @@ export default function PaginationControls({
         </div>
       </div>
 
-      <nav className="flex items-center justify-between gap-2" aria-label={ariaLabel}>
+      <nav
+        className="flex items-center justify-between gap-2"
+        aria-label={ariaLabel}
+      >
         <div className="flex items-center gap-1">
           <Button
             variant="outline"
@@ -84,7 +103,11 @@ export default function PaginationControls({
         <div className="hidden sm:flex items-center gap-1">
           {range.map((item, idx) =>
             item === "…" ? (
-              <span key={`ellipsis-${idx}`} className="px-2 text-muted-foreground" aria-hidden>
+              <span
+                key={`ellipsis-${idx}`}
+                className="px-2 text-muted-foreground"
+                aria-hidden
+              >
                 …
               </span>
             ) : (
@@ -96,7 +119,7 @@ export default function PaginationControls({
                   "h-8 w-8 p-0",
                   page === item
                     ? "bg-cyan-500 hover:bg-cyan-600 text-white shadow-md"
-                    : "hover:bg-gray-100"
+                    : "hover:bg-gray-100",
                 )}
                 aria-current={page === item ? "page" : undefined}
                 aria-label={`Aller à la page ${item}`}
@@ -104,11 +127,14 @@ export default function PaginationControls({
               >
                 {item}
               </Button>
-            )
+            ),
           )}
         </div>
 
-        <div className="sm:hidden px-4 py-1 bg-[white] border-cyan-500 border rounded-xl text-sm font-medium" aria-live="polite">
+        <div
+          className="sm:hidden px-4 py-1 bg-[white] border-cyan-500 border rounded-xl text-sm font-medium"
+          aria-live="polite"
+        >
           {page}/{totalPages}
         </div>
