+++ src/components/admin/stations/StationsTable.tsx
@@ -50,6 +50,7 @@ import { cn } from "@/lib/utils";
 import CardStationStats from "./CardStats";
 import { Alert, AlertDescription, AlertTitle } from "@/components/ui/alert";
 import ActionBar from "./ActionBar";
+import { usePagination } from "@/hooks/usePagination";
 import PaginationControls from "../users/Pagination";
 import UpdateStationForm from "./UpdateStationForm";
 import DeleteStationAction from "./DialogConfirmationDeleteStation";
@@ -81,42 +82,6 @@ type ConfirmDialogState = {
 
 const ITEMS_PER_PAGE = 10;
 
-function usePagination(
-  totalItems: number,
-  itemsPerPage: number = ITEMS_PER_PAGE
-) {
-  const [page, setPage] = useState(1);
-  const totalPages = Math.max(1, Math.ceil(totalItems / itemsPerPage));
-
-  const goToNextPage = useCallback(() => {
-    setPage((prev) => Math.min(prev + 1, totalPages));
-  }, [totalPages]);
-
-  const goToPrevPage = useCallback(() => {
-    setPage((prev) => Math.max(prev - 1, 1));
-  }, []);
-
-  const goToPage = useCallback(
-    (pageNum: number) => {
-      setPage(Math.max(1, Math.min(pageNum, totalPages)));
-    },
-    [totalPages]
-  );
-
-  const resetPage = useCallback(() => setPage(1), []);
-
-  return {
-    page,
-    totalPages,
-    itemsPerPage,
-    goToNextPage,
-    goToPrevPage,
-    goToPage,
-    resetPage,
-    isFirstPage: page === 1,
-    isLastPage: page === totalPages,
-  };
-}
 
 function StationTableRow({
   station,
