+++ src/components/admin/users/UsersTable.tsx
@@ -44,6 +44,7 @@ import { useTranslations } from "next-intl";
 import CardUserStats from "./CardStats";
 import { Alert, AlertDescription, AlertTitle } from "@/components/ui/alert";
 import ActionBar from "./ActionBar";
+import { usePagination } from "@/hooks/usePagination";
 import PaginationControls from "./Pagination";
 import ResetPasswordAction from "./DialogConfirmationResetsPassword";
 import DeleteUserAction from "./DialogConfirmationDeleteUser";
@@ -83,42 +84,6 @@ function getCurrentUserIdFromCookie(): number | null {
   }
 }
 
-function usePagination(
-  totalItems: number,
-  itemsPerPage: number = ITEMS_PER_PAGE
-) {
-  const [page, setPage] = useState(1);
-  const totalPages = Math.max(1, Math.ceil(totalItems / itemsPerPage));
-
-  const goToNextPage = useCallback(() => {
-    setPage((prev) => Math.min(prev + 1, totalPages));
-  }, [totalPages]);
-
-  const goToPrevPage = useCallback(() => {
-    setPage((prev) => Math.max(prev - 1, 1));
-  }, []);
-
-  const goToPage = useCallback(
-    (pageNum: number) => {
-      setPage(Math.max(1, Math.min(pageNum, totalPages)));
-    },
-    [totalPages]
-  );
-
-  const resetPage = useCallback(() => setPage(1), []);
-
-  return {
-    page,
-    totalPages,
-    itemsPerPage,
-    goToNextPage,
-    goToPrevPage,
-    goToPage,
-    resetPage,
-    isFirstPage: page === 1,
-    isLastPage: page === totalPages,
-  };
-}
 
 function RoleBadge({ isAdmin }: { isAdmin: boolean }) {
   const t = useTranslations();
